+++ client/src/main.tsx
@@ -6,7 +6,6 @@ import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 
 import ENV from "@app/env";
 import App from "@app/App";
-import reportWebVitals from "@app/reportWebVitals";
 import "@app/dayjs";
 import { OidcProvider } from "@app/components/OidcProvider";
 
@@ -39,8 +38,3 @@ const renderApp = () => {
 //   renderApp();
 // }
 renderApp();
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
