+++ package.json
@@ -48,8 +48,7 @@
     "prettier": "^3.0.2",
     "rimraf": "^5.0.7",
     "rollup": "^3.29.4",
-    "rollup-plugin-copy": "^3.5.0",
-    "vite-tsconfig-paths": "^5.1.4"
+    "rollup-plugin-copy": "^3.5.0"
   },
   "peerDependencies": {
     "react": "^18.2.0",
